Autodiscover project files and accept additional files
