+++ src/constants.ts
@@ -28,4 +28,6 @@ export const SITE_ENV_VAR = 'DD_SITE'
 // Flare constants
 export const FLARE_OUTPUT_DIRECTORY = '.datadog-ci'
 export const LOGS_DIRECTORY = 'logs'
+export const PROJECT_FILES_DIRECTORY = 'project_files'
+export const ADDITIONAL_FILES_DIRECTORY = 'additional_files'
 export const FLARE_ENDPOINT_PATH = '/api/ui/support/serverless/flare'
