+++ src/helpers/__tests__/prompt.test.ts
@@ -1,7 +1,7 @@
 jest.mock('inquirer')
 import {prompt} from 'inquirer'
 
-import {confirmationQuestion, requestConfirmation} from '../prompt'
+import {confirmationQuestion, requestConfirmation, requestFilePath} from '../prompt'
 
 describe('prompt', () => {
   describe('confirmationQuestion', () => {
@@ -35,6 +35,29 @@ describe('prompt', () => {
         }
       }
       expect(error?.message).toBe("Couldn't receive confirmation. Unexpected error")
+    })
+  })
+
+  describe('requestFilePath', () => {
+    const mockFilePath = '/Users/username/project/test.ts'
+    test('returns the selected file path', async () => {
+      ;(prompt as any).mockImplementation(() => Promise.resolve({filePath: mockFilePath}))
+
+      const selectedPath = await requestFilePath()
+      expect(mockFilePath).toBe(selectedPath)
+    })
+
+    test('throws error when something unexpected happens while prompting', async () => {
+      ;(prompt as any).mockImplementation(() => Promise.reject(new Error('Unexpected error')))
+      let error
+      try {
+        await requestFilePath()
+      } catch (e) {
+        if (e instanceof Error) {
+          error = e
+        }
+      }
+      expect(error?.message).toBe("Couldn't receive file path. Unexpected error")
     })
   })
 })
