+++ src/helpers/prompt.ts
@@ -22,3 +22,21 @@ export const requestConfirmation = async (message: string, defaultValue = true) 
     }
   }
 }
+
+export const requestFilePath = async () => {
+  try {
+    const filePathAnswer = await inquirer.prompt([
+      {
+        type: 'input',
+        name: 'filePath',
+        message: 'Please enter a file path, or press Enter to finish:',
+      },
+    ])
+
+    return filePathAnswer.filePath
+  } catch (e) {
+    if (e instanceof Error) {
+      throw Error(`Couldn't receive file path. ${e.message}`)
+    }
+  }
+}
